requirements.txt
# Enhanced Document AI Multi-Agent Assistant - Requirements
# Core dependencies for the enhanced version with Guardrails AI, testing, and resilience

# Web Interface
gradio>=4.0.0,<5.0.0

# AI/ML Libraries
google-generativeai>=0.3.0,<1.0.0

# Guardrails AI - Enhanced Safety
guardrails-ai>=0.4.0,<1.0.0
# Guardrails Hub validators (install as needed)
# pip install guardrails-ai[all] for all validators

# HTTP and API
requests>=2.28.0,<3.0.0
urllib3>=1.26.0,<2.0.0

# Data Processing
PyPDF2>=3.0.0,<4.0.0
python-docx>=0.8.11,<1.0.0
beautifulsoup4>=4.11.0,<5.0.0

# Environment and Configuration
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0,<7.0

# Testing Framework
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
pytest-mock>=3.10.0,<4.0.0
pytest-asyncio>=0.21.0,<1.0.0

# Code Quality
black>=22.0.0,<24.0.0
flake8>=5.0.0,<7.0.0
isort>=5.10.0,<6.0.0
mypy>=1.0.0,<2.0.0

# Security and Safety (Enhanced)
cryptography>=3.4.8,<42.0.0
hashlib-compat

# For Guardrails AI dependencies
pydantic>=2.0.0,<3.0.0
rich>=13.0.0,<14.0.0
typer>=0.9.0,<1.0.0

# Natural Language Processing (for Guardrails)
spacy>=3.4.0,<4.0.0
transformers>=4.21.0,<5.0.0
torch>=1.12.0,<3.0.0

# Utilities
click>=8.0.0,<9.0.0
colorama>=0.4.4,<1.0.0
tqdm>=4.64.0,<5.0.0

# Logging and Monitoring
structlog>=22.0.0,<24.0.0
python-json-logger>=2.0.0,<3.0.0

# Date and Time
python-dateutil>=2.8.0,<3.0.0

# File System
watchdog>=2.1.0,<4.0.0
pathlib2>=2.3.0,<3.0.0;python_version<"3.4"

# Memory Management
psutil>=5.9.0,<6.0.0
memory-profiler>=0.60.0,<1.0.0

# Development Tools (Optional)
pre-commit>=2.20.0,<4.0.0
coverage>=6.0.0,<8.0.0

# Documentation
mkdocs>=1.4.0,<2.0.0
mkdocs-material>=8.0.0,<10.0.0

# Performance
cachetools>=5.0.0,<6.0.0
lru-dict>=1.1.0,<2.0.0

# Async Support
asyncio>=3.4.3;python_version<"3.7"
aiofiles>=22.0.0,<24.0.0

# JSON Processing
orjson>=3.8.0,<4.0.0

# Regular Expressions (Enhanced)
regex>=2022.0.0,<2025.0.0

# HTTP Server (for production deployment)
uvicorn>=0.20.0,<1.0.0
fastapi>=0.95.0,<1.0.0

# Database (for state persistence - optional)
sqlite3-to-mysql>=1.4.0,<2.0.0

# Compression
gzip>=1.0.0
zstandard>=0.19.0,<1.0.0

# Image Processing (for future PDF image extraction)
Pillow>=9.0.0,<11.0.0

# Scientific Computing (for advanced text analysis and Guardrails)
numpy>=1.21.0,<2.0.0
scipy>=1.9.0,<2.0.0

# Text Processing (required for Guardrails AI)
nltk>=3.8.0,<4.0.0

# Machine Learning (for Guardrails validators)
scikit-learn>=1.1.0,<2.0.0

# Compatibility
typing-extensions>=4.0.0,<5.0.0
importlib-metadata>=4.0.0,<7.0.0;python_version<"3.8"

# Error Tracking (Optional)
sentry-sdk>=1.15.0,<2.0.0

# Additional Guardrails AI Hub validators (optional but recommended)
# Uncomment specific validators as needed:
# guardrails[pii]>=0.4.0  # For PII detection
# guardrails[toxic]>=0.4.0  # For toxicity detection 
# guardrails[competitors]>=0.4.0  # For competitor mention detection

# Note: Some Guardrails validators may require additional model downloads
# Run after installation: python -c "import spacy; spacy.cli.download('en_core_web_sm')"


# Web Interface
gradio>=4.0.0,<5.0.0

# AI/ML Libraries
google-generativeai>=0.3.0,<1.0.0

# HTTP and API
requests>=2.28.0,<3.0.0
urllib3>=1.26.0,<2.0.0

# Data Processing
PyPDF2>=3.0.0,<4.0.0
python-docx>=0.8.11,<1.0.0
beautifulsoup4>=4.11.0,<5.0.0

# Environment and Configuration
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0,<7.0

# Testing Framework
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
pytest-mock>=3.10.0,<4.0.0
pytest-asyncio>=0.21.0,<1.0.0

# Code Quality
black>=22.0.0,<24.0.0
flake8>=5.0.0,<7.0.0
isort>=5.10.0,<6.0.0
mypy>=1.0.0,<2.0.0

# Security and Safety
cryptography>=3.4.8,<42.0.0
hashlib-compat>=1.0.0,<2.0.0

# Utilities
click>=8.0.0,<9.0.0
colorama>=0.4.4,<1.0.0
tqdm>=4.64.0,<5.0.0

# Logging and Monitoring
structlog>=22.0.0,<24.0.0
python-json-logger>=2.0.0,<3.0.0

# Date and Time
python-dateutil>=2.8.0,<3.0.0

# File System
watchdog>=2.1.0,<4.0.0
pathlib2>=2.3.0,<3.0.0;python_version<"3.4"

# Memory Management
psutil>=5.9.0,<6.0.0
memory-profiler>=0.60.0,<1.0.0

# Development Tools (Optional)
pre-commit>=2.20.0,<4.0.0
coverage>=6.0.0,<8.0.0

# Documentation
mkdocs>=1.4.0,<2.0.0
mkdocs-material>=8.0.0,<10.0.0

# Performance
cachetools>=5.0.0,<6.0.0
lru-dict>=1.1.0,<2.0.0

# Async Support
asyncio>=3.4.3;python_version<"3.7"
aiofiles>=22.0.0,<24.0.0

# JSON Processing
orjson>=3.8.0,<4.0.0

# Regular Expressions (Enhanced)
regex>=2022.0.0,<2025.0.0

# HTTP Server (for production deployment)
uvicorn>=0.20.0,<1.0.0
fastapi>=0.95.0,<1.0.0

# Database (for state persistence - optional)
sqlite3-to-mysql>=1.4.0,<2.0.0

# Compression
gzip>=1.0.0
zstandard>=0.19.0,<1.0.0

# Image Processing (for future PDF image extraction)
Pillow>=9.0.0,<11.0.0

# Scientific Computing (for advanced text analysis)
numpy>=1.21.0,<2.0.0
scipy>=1.9.0,<2.0.0

# Text Processing
nltk>=3.8.0,<4.0.0
spacy>=3.4.0,<4.0.0

# Compatibility
typing-extensions>=4.0.0,<5.0.0
importlib-metadata>=4.0.0,<7.0.0;python_version<"3.8"

# Error Tracking (Optional)
sentry-sdk>=1.15.0,<2.0.0
